add keyboard navigation to all routes in compactView preview tree; TOO SLOW
