+++ packages/cli/src/pages/previews/[previewClass]/[previewFunction].tsx
@@ -1,18 +1,14 @@
-import React, { useCallback, useState } from "react";
-import { useRouter } from "next/router";
-import Header from "../../../components/Header";
-import HotIFrame from "../../../components/HotIFrame";
-import MjmlErrors from "../../../components/MjmlErrors";
+import React from "react";
 import { GetStaticProps, NextPage } from "next";
-import { hotkeysMap } from "../../../components/hooks/usePreviewHotkeys";
-import useLiveReload from "../../../components/hooks/useLiveReload";
+
 import { render } from "../../../util/mjml";
 import {
   getPreviewComponent,
   previewTree,
 } from "../../../util/moduleManifestUtil";
-import IndexPane from "../../../components/IndexPane/IndexPane";
-import CircleLoader from "../../../components/CircleLoader";
+import PreviewViewer, {
+  PreviewViewerProps,
+} from "../../../components/PreviewViewer";
 
 type Params = { previewClass: string; previewFunction: string };
 
@@ -52,171 +48,8 @@ export const getStaticProps: GetStaticProps = async (context) => {
   };
 };
 
-type Data = {
-  preview: ShowPreviewResponseBody;
-  previews: [string, string[]][];
-};
-
-type PreviewProps = {
-  initialData: Data;
-};
-
-async function fetchJson(url: string) {
-  const response = await fetch(url, {
-    headers: { "Content-Type": "application/json" },
-  });
-  return await response.json();
-}
-
-const Preview: NextPage<PreviewProps> = ({ initialData }) => {
-  const router = useRouter();
-  const { previewClass, previewFunction } = router.query;
-  const [viewMode, setViewMode] = useState<ViewMode>("desktop");
-  const [data, setData] = useState<Data>(initialData);
-  const [fetching, setFetching] = useState(false);
-
-  const fetchData = useCallback(async () => {
-    if (!previewClass || !previewFunction) return;
-    setFetching(true);
-    const json = await fetchJson(
-      `/api/previews/${previewClass}/${previewFunction}`
-    );
-
-    setData({
-      preview: json,
-      previews: json.previews,
-    });
-    setFetching(false);
-  }, [setData, previewClass, previewFunction]);
-
-  useLiveReload(fetchData);
-
-  const { preview, previews } = data || { preview: null, previews: [] };
-
-  return (
-    <div>
-      <Header
-        title={`${previewClass} - ${previewFunction}`}
-        previewClass={previewClass as string}
-        previewFunction={previewFunction as string}
-        viewMode={viewMode}
-        setViewMode={setViewMode}
-        helpContent={
-          <>
-            <div className="title">Hotkeys</div>
-            <div className="hotkey">
-              <span className="character">{hotkeysMap.showPreviews}</span>
-              <span className="description">Jump to previews</span>
-            </div>
-            <div className="hotkey">
-              <span className="character">{hotkeysMap.viewModeNext}</span>
-              <span className="description">Next view mode</span>
-            </div>
-            <div className="hotkey">
-              <span className="character">{hotkeysMap.viewModePrevious}</span>
-              <span className="description">Previous view mode</span>
-            </div>
-            <div className="hotkey">
-              <span className="character">{hotkeysMap.viewModeDesktop}</span>
-              <span className="description">Desktop view</span>
-            </div>
-            <div className="hotkey">
-              <span className="character">{hotkeysMap.viewModeMobile}</span>
-              <span className="description">Mobile view</span>
-            </div>
-            <div className="hotkey">
-              <span className="character">{hotkeysMap.viewModeHTML}</span>
-              <span className="description">HTML view</span>
-            </div>
-          </>
-        }
-      />
-
-      <div>
-        <div className="left-pane">
-          <IndexPane previews={previews} />
-        </div>
-        <div className="right-pane">
-          {!!preview?.errors.length && <MjmlErrors errors={preview?.errors} />}
-          {preview?.html && !preview?.errors.length && (
-            <HotIFrame
-              srcDoc={preview.html}
-              viewMode={viewMode}
-              setViewMode={setViewMode}
-            />
-          )}
-          {fetching && (
-            <div className="loader-position">
-              <CircleLoader />
-            </div>
-          )}
-        </div>
-      </div>
-
-      <style jsx>{`
-        .left-pane,
-        .right-pane {
-          position: absolute;
-          overflow: scroll;
-          top: 65px;
-          bottom: 0;
-        }
-        .left-pane {
-          width: 300px;
-          left: 0;
-        }
-        .right-pane {
-          right: 0;
-          left: 300px;
-          width: calc(100vw - 300px);
-        }
-        .title {
-          padding-bottom: 4px;
-        }
-        .title,
-        .character {
-          text-transform: uppercase;
-          font-size: 10px;
-          line-height: 100%;
-        }
-        .hotkey {
-          font-size: 12px;
-          margin: 12px 24px 0 0;
-        }
-        .character {
-          color: #bbb;
-          width: 18px;
-          height: 18px;
-          border: solid 1px #999;
-          border-radius: 2px;
-          text-align: center;
-          margin-right: 8px;
-          display: inline-block;
-          line-height: 170%;
-        }
-        .description {
-          position: relative;
-          top: 1.25px;
-        }
-        .fetch-indicator {
-          z-index: 9999;
-          height: 100px;
-          width: 100px;
-          top: 0;
-          left: 0;
-          right: 0;
-          bottom: 0;
-          font-size: 48px;
-          background-color: pink;
-        }
-        .loader-position {
-          position: absolute;
-          bottom: 24px;
-          right: 24px;
-        }
-      `}</style>
-    </div>
-  );
+const Preview: NextPage<PreviewViewerProps> = ({ initialData }) => {
+  return <PreviewViewer initialData={initialData} />;
 };
 
 export default Preview;
